# pm: Optimize power state selection with early exit when no state is available