Issues found in power management:

### Lack of early exit when there are no power states available

When low latency requirement is set (e.g. 0us) then `pm_suspend` does not perform faster as it still iterates over all power modes only to find that there is no power mode that fulfills the requirement. When all power modes are locked `pm_suspend` behavior is the same as when latency requirement is set. In both cases user sets those to get higher performance so ideally, `pm_suspend` should be limited.

Added `pm_policy_state_any_active` which is using bitmasks to determine if any state is available. It is used early in the `pm_suspend` and allows early exit. Added `pm_policy_state_is_available` which checks if state is locked or latency requirement is not met.

### Default algorithm for picking power state iterates always from the last power state

Algorithm for picking power state iterates always from the last power state (the longest sleep) and because of that `pm_suspend` is the longest for the shortest (most common) sleep periods.

Order is reverted.

### Performance results

Timings were measured on nrf54h20 with 2 power states. Code includes other proposed improvements: #88144 and #88146. It measures the time between entering `pm_suspend` and selecting the power state or early exit. Results show significant improvement for case where no state is available and 33% improvement for shortest sleep.

| Operation  | Before | After |
| ------------- | ------------- | --- |
| Short timeout, stays in ACTIVE  | 305  | 200 |
| 1st power mode  | 309  | 252 |
| 2st power mode | 261 | 284 |
| Latency requirement set to 0 | 289 | 57 |
| All modes locked | 277 | 57 |

Results show that high performance case is significantly improved.
Short sleep times is slightly improved.